INC-71 refactor: move Google auth logic to custom hook
